feat(toolbar): Display toolbar/navbar links as per design (#184)
* feat(toolbar): Display toolbar/navbar links as per design

This change makes it so the toolbar displyed via the app is per the design. A change has been made
to the NavLink component to make it render Router aware anchor Links, or external anchor links
depending on the string provided via the "to" prop. Because clicking these external links takes you
away from the site, they can never receive the active style. It would be possible to make these
links open in a new tab, if we decide to do that in future.

fix #144