fix(#590): Last two bits for this issue (#613)
* fix(#590): Add taxonomy type to the results page

Also simplified the list code as the taxonomy was available as a prop, there was no need to get it from the query

* chore(#590): Setup Jest

* chore(#590): Factor out article summary image panel

* feat(#590): Allow ArticleVisualSummary to be a link or not

* fix(#590): Tighter image dimensions

* feat(#590): Reuse article visual summary panel on home page

* fix(#590): Ensure label only one line deep

* fix(#590): Vertical spacing (keep this in the layout parent)

* fix(#590): Test needed wrapper to work

* fix(#590): xit test (see description)

Test seemed to be testing something the component doesn't handle - namely that an unauthenticated user gets a certain response. The code is checking the user's role and protecting certain routes based on that, not returning different things based on authentication state.

* fix(#592): Further Reading layout

* fix(#590): Various grid things to remove bloat and re-use article summary layout

* feat(#590): Improved HTML and wrote tests

* fix: Edge flat fatal error

'Usage of Array.flat() was breaking Edge as it'\''s not supported by that browser. Intention was to replace with Lodash.flatten() however the line wasn'\''t actually assgning the return value to anything. As flat doesn'\''t operate on the array in place this line was having no effect and therefore was removed.

* fix(#590): Text styles for summary labels reverted

* fix(#590): Grid min size fixed rather than content driven

* fix(#590): Text component should still be a p