feat(#610): Add a modal with data for KeyInfo in the assessment
Along the way also makes the ContentModal more re-usable by letting the width and open state be configurable, as well as allowing an optional `subTitle` prop to be added.
Uses markdown in separate files rather than overloading Matrix JSON and also as the same text will be used for both matrices.