fix: 🐛 autosave conflict on criteria parts loss of data
When a new criteria parts page was open in 2 browsers, the autosave with
watch functionality would conflict with undefined results. The core of
the problem was that the component state was not updating when the new
part was saved in the other browser, resulting in the saving in the
original browswer also doing an insert instead of an update.